# docs/apifox/auth-api.openapi.yaml
openapi: 3.0.3
info:
  title: 智能抄表系统后端接口
  description: 当前认证、用户管理、表具管理、居民表具绑定、设备、告警、抄表任务及阿里云 OSS 抄表图片模块的 Apifox 测试文档。
  version: 2.8.0
servers:
  - url: http://localhost:8070
    description: 本地开发环境
...
paths:
...
  /api/v1/admin/meters/{meterId}:
    parameters:
      - $ref: '#/components/parameters/MeterIdPath'
    get:
      tags:
        - 表具管理
      summary: 查询表具详情
      description: 查询未删除表具的完整档案。仅 ADMIN 可访问。
      operationId: getMeter
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterDetailSuccessResult'
        '400':
          description: meterId 类型不正确
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '401':
          description: 未登录或登录会话已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '403':
          description: 当前用户没有 ADMIN 角色
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '404':
          description: 表具不存在或已经逻辑删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'

    put:
      tags:
        - 表具管理
      summary: 修改表具基础资料
      description: 修改名称、类型、显示类型、单位、读数位数、初始读数、安装日期和备注；不修改表具编号和状态。使用 version 做乐观锁校验。仅 ADMIN 可访问。
      operationId: updateMeter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeterDTO'
            example:
              meterName: 1号电表
              meterType: ELECTRIC
              displayType: LCD
              unit: kWh
              integerDigits: 8
              decimalDigits: 2
              initialReading: 98.76
              installedAt: '2026-09-17'
              remark: 更换为液晶表
              version: 0
      responses:
        '200':
          description: 修改成功，返回递增后的版本号
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterVersionSuccessResult'
        '400':
          description: 字段校验失败、业务规则不满足或已报废表具禁止修改
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResult'
                  - $ref: '#/components/schemas/ErrorResult'
        '401':
          description: 未登录或登录会话已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '403':
          description: 当前用户没有 ADMIN 角色
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '404':
          description: 表具不存在或已经逻辑删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '409':
          description: version 已过期，请重新查询详情后提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              example:
                code: VERSION_CONFLICT
                message: 表具数据已被其他请求修改，请刷新后重试
                data: null
                traceId: 01JEXAMPLETRACEID
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'

    delete:
      tags:
        - 表具管理
      summary: 删除表具
      description: 仅允许逻辑删除停用且没有居民、设备绑定关系的表具；version 必须与数据库一致。删除后普通查询不可见，数据库记录保留。仅 ADMIN 可访问。
      operationId: deleteMeter
      security:
        - bearerAuth: []
      parameters:
        - name: version
          in: query
          required: true
          description: 客户端最近一次查询到的数据版本
          schema:
            type: integer
            minimum: 0
          example: 2
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptySuccessResult'
        '400':
          description: 未传 version、version 类型错误或表具状态不是停用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '401':
          description: 未登录或登录会话已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '403':
          description: 当前用户没有 ADMIN 角色
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '404':
          description: 表具不存在或已经逻辑删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '409':
          description: version 已过期，或者表具仍绑定居民或设备
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
...
  /api/v1/meter-reader/tasks/{taskId}/images:
    post:
      tags: [抄表图片]
      summary: 抄表员上传任务图片
      description: 仅允许给本人 PROCESSING 任务上传 JPEG、PNG 或 WebP；同一 X-Idempotency-Key 重试返回同一图片。
      operationId: uploadReaderMeterImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                imageType:
                  $ref: '#/components/schemas/MeterImageType'
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 上传成功或命中幂等记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageItemSuccessResult'
        '400':
          description: 文件为空、格式不支持、文件过大或请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '404':
          description: 任务不存在或不属于当前抄表员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '409':
          description: 任务状态变化、达到图片上限或幂等标识冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '502':
          description: OSS 未配置或 OSS 操作失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
    get:
      tags: [抄表图片]
      summary: 查询本人任务图片
      operationId: listReaderMeterImages
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeterImageListSuccessResult'
  /api/v1/meter-reader/tasks/{taskId}/images/{imageId}:
    delete:
      tags: [抄表图片]
      summary: 抄表员删除误传图片
      description: 仅允许软删除本人 PROCESSING 任务下尚未绑定结果的图片；OSS 删除失败时由定时任务补偿。
      operationId: deleteReaderMeterImage
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/ImageIdPath'
        - name: version
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: 删除请求处理成功
        '404':
          description: 图片不存在或不属于当前抄表员
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
        '409':
          description: 图片已绑定结果、任务状态变化或版本冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    deviceNoHeader:
      type: apiKey
      in: header
      name: X-Device-No
      description: 管理员创建设备时指定的全局唯一设备编号
    deviceSecretHeader:
      type: apiKey
      in: header
      name: X-Device-Secret
      description: 创建设备或重置密钥时仅返回一次的设备明文密钥
  parameters:
...
    MeterIdPath:
      name: meterId
      in: path
      required: true
      description: 目标表具 ID
      schema:
        type: integer
        format: int64
      example: 1
...
    TaskIdPath:
      name: taskId
      in: path
      required: true
      description: 抄表任务 ID
      schema:
        type: integer
        format: int64
        minimum: 1
    ImageIdPath:
      name: imageId
      in: path
      required: true
      description: 抄表图片元数据 ID
      schema:
        type: integer
        format: int64
        minimum: 1
    IdempotencyKeyHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      description: 每次新上传生成一个唯一值；同一请求重试时保持不变，最大64个字符
      schema:
        type: string
        minLength: 1
        maxLength: 64
      example: 550e8400-e29b-41d4-a716-446655440000
  schemas:
...
    MeterType:
      type: string
      enum:
        - WATER
        - ELECTRIC
        - GAS
      description: 表具测量类型：水表、电表、燃气表
    MeterDisplayType:
      type: string
      enum:
        - LCD
        - MECHANICAL_ROLLER
      description: 表具显示类型：液晶显示、机械滚轮
    UpdateMeterDTO:
      type: object
      required:
        - meterName
        - meterType
        - displayType
        - unit
        - integerDigits
        - decimalDigits
        - initialReading
        - version
      properties:
        meterName:
          type: string
          minLength: 1
          maxLength: 64
          description: 表具名称
        meterType:
          $ref: '#/components/schemas/MeterType'
        displayType:
          $ref: '#/components/schemas/MeterDisplayType'
        unit:
          type: string
          minLength: 1
          maxLength: 16
          description: WATER、GAS 必须为 m³，ELECTRIC 必须为 kWh
        integerDigits:
          type: integer
          minimum: 1
          maximum: 12
        decimalDigits:
          type: integer
          minimum: 0
          maximum: 3
        initialReading:
          type: number
          minimum: 0
          multipleOf: 0.001
        installedAt:
          type: string
          format: date
          nullable: true
          description: 安装日期，不能晚于今天；传 null 可清空
        remark:
          type: string
          maxLength: 500
          nullable: true
          description: 备注；传 null 或纯空白可清空
        version:
          type: integer
          minimum: 0
          description: 客户端最近一次查询到的数据版本
...
    MeterDetailVO:
      type: object
      required:
        - meterId
        - meterNo
        - meterName
        - meterType
        - displayType
        - unit
        - integerDigits
        - decimalDigits
        - initialReading
        - status
        - version
        - createdBy
        - updatedBy
        - createdAt
        - updatedAt
      properties:
        meterId:
          type: integer
          format: int64
        meterNo:
          type: string
        meterName:
          type: string
        meterType:
          $ref: '#/components/schemas/MeterType'
        displayType:
          $ref: '#/components/schemas/MeterDisplayType'
        unit:
          type: string
        integerDigits:
          type: integer
        decimalDigits:
          type: integer
        initialReading:
          type: number
        installedAt:
          type: string
          format: date
          nullable: true
        status:
          type: integer
          enum:
            - 0
            - 1
            - 2
            - 3
          description: 0 停用，1 正常，2 维护中，3 已报废
        version:
          type: integer
        remark:
          type: string
          nullable: true
        createdBy:
          type: integer
          format: int64
          description: 创建该表具的管理员 ID
        updatedBy:
          type: integer
          format: int64
          description: 最后修改该表具的管理员 ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MeterDetailSuccessResult:
      type: object
      required:
        - code
        - message
        - data
        - traceId
      properties:
        code:
          type: string
          enum:
            - OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/MeterDetailVO'
        traceId:
          type: string
    MeterVersionVO:
      type: object
      required:
        - meterId
        - version
      properties:
        meterId:
          type: integer
          format: int64
          example: 1
        version:
          type: integer
          description: 操作成功后的最新乐观锁版本
          example: 1
    MeterVersionSuccessResult:
      type: object
      required:
        - code
        - message
        - data
        - traceId
      properties:
        code:
          type: string
          enum:
            - OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/MeterVersionVO'
        traceId:
          type: string
...
    MeterImageType:
      type: string
      enum: [ORIGINAL, AI_ANNOTATED, ENVIRONMENT]
      default: ORIGINAL
      description: ORIGINAL 原始表具图片；AI_ANNOTATED AI 标注图片；ENVIRONMENT 现场环境图片
    MeterImageStatus:
      type: string
      enum: [VALID, INVALID]
      description: 图片业务有效性状态
    MeterImageStorageStatus:
      type: string
      enum: [STORED, DELETE_PENDING, DELETED]
      description: OSS 对象存储生命周期状态
...
    MeterImageItemVO:
      type: object
      required:
        - imageId
        - taskId
        - imageType
        - originalName
        - contentType
        - fileSize
        - imageStatus
        - storageStatus
        - version
        - createdAt
      properties:
        imageId: {type: integer, format: int64}
        taskId: {type: integer, format: int64}
        imageType: {$ref: '#/components/schemas/MeterImageType'}
        imageTypeName: {type: string}
        originalName: {type: string}
        contentType: {type: string, enum: [image/jpeg, image/png, image/webp]}
        fileSize: {type: integer, format: int64, description: 字节数}
        imageWidth: {type: integer, nullable: true}
        imageHeight: {type: integer, nullable: true}
        imageStatus: {$ref: '#/components/schemas/MeterImageStatus'}
        imageStatusName: {type: string}
        storageStatus: {$ref: '#/components/schemas/MeterImageStorageStatus'}
        resultId: {type: integer, format: int64, nullable: true}
        version: {type: integer, minimum: 0}
        createdAt: {type: string, format: date-time}
...
    MeterImageItemSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data: {$ref: '#/components/schemas/MeterImageItemVO'}
        traceId: {type: string}
    MeterImageListSuccessResult:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: {type: string, enum: [OK]}
        message: {type: string, example: success}
        data:
          type: array
          items: {$ref: '#/components/schemas/MeterImageItemVO'}
        traceId: {type: string}
...
    EmptySuccessResult:
      type: object
      required:
        - code
        - message
        - data
        - traceId
      properties:
        code:
          type: string
          enum:
            - OK
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true
          description: 当前操作没有业务响应数据，固定为 null
        traceId:
          type: string
          description: 本次请求的追踪标识
...
    ValidationErrorResult:
      type: object
      required:
        - code
        - message
        - data
        - traceId
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
        message:
          type: string
          example: 请求参数不正确
        data:
          type: object
          additionalProperties:
            type: string
        traceId:
          type: string
    ErrorResult:
      type: object
      required:
        - code
        - message
        - data
        - traceId
      properties:
        code:
          type: string
          example: AUTH_FAILED
        message:
          type: string
          example: 用户名或密码错误
        data:
          type: object
          nullable: true
          description: 错误响应时为 null
        traceId:
          type: string